requirements.txt
bandit==1.8.3
botocore==1.38.23
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
docutils==0.19
filelock==3.18.0
flake8==7.2.0
flake8-quotes==3.4.0
fsspec==2025.5.1
idna==3.10
jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
loguru==0.7.3
markdown-it-py==3.0.0
markupsafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.6
pandas==2.2.3
pbr==6.1.1
pip==25.1.1
pyasn1==0.6.1
pycodestyle==2.13.0
pyflakes==3.3.2
pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytest==7.4.4
pytz==2025.2
pyyaml==6.0.2
regex==2024.11.6
#remla2025-team19-lib-ml @ git+https://github.com/remla2025-team19/lib-ml.git@c549aec09929fc0a0d20fac439efd363e84cbbca
requests==2.32.3
rich==14.0.0
rsa==4.7.2
ruff==0.11.11
s3transfer==0.13.0
scikit-learn==1.6.1
scipy==1.15.3
setuptools==80.8.0
shellingham==1.5.4
six==1.17.0
stevedore==5.4.1
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.7.0
tqdm==4.67.1
typer==0.15.3
typing-extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
-e .

pandas>=1.5.0
scikit-learn>=1.0.0
pylint>=2.17
flake8>=6.1
bandit>=1.7
remla2025-team19-lib-ml @ git+https://github.com/remla2025-team19/lib-ml.git@v1.0.0